Make MemoryDataStore.save async.
The save method renders all the html and json to disk which is a slow
operation that can would block the event loop. We move both to a thread
pool which works on a clone of the MemoryDataStore.

The save method also makes network I/O to fetch node details of nodes.
This was making clocking network calls with requests. This now uses
aiohttp which is async and will not block the event loop.