Resolve non numerical values for temperature's
Specifically with a users with a Wio Tracker 1110 Dev Kit for Meshtastic produce sometimes the value NAN for temperature. This appears to be the only device I've seen it from but this change would resolve any future issues. When meshinfo sees this value it crashes the app. Has been stable for a week on my states current setup and the node continues to run.